Revert ".emacs: optimize `unpack` macro"
This reverts commit 167cda5420edfab0cebf3d6f5f7f65ae452c8959.

This optimization broke unpacking in tricky ways.

The mose obvious case is `consult-fd`, which just
fails with an `overlayp` failure on a nil. The
unpacking that I have in the window state code
starts to fail after I evaluate

    (unpack (a b c d) '(1 2 3 4))

so clearly I'm somehow leaking mutable state
between macro calls, but I am done messing
with it for now.