.emacs: first pass at pop-command-in-Eshell-window
The biggest advantage of doing this in Eshell
rather than a raw Eat buffer is integration:
`with-editor` works in Eshell but not in Eat
for me (I remember concluding that I fixed it
but I guess not), I get to keep historical
command output between runs until I choose to
delete it, and it also solves the terminal
sizing issue automatically.

The biggest downside right now with this
implementation is that it cannot pass through
`same-window-prefix`, which is yet another
thing I would have to figure out how to do.

I think ultimately the right way to do this
is to manually implement something like how
`eshell-command` is implemented.