.emacs: finally, adopting a little Magit
One of the reasons I kept not adopting Magit
was that it felt slow - turns out this was
mostly initial package load slowness, which
does not happen on subsequent runs. Making a
commit is still perceptibly slower on my
Android phone on Termux than without Magit,
but not terribly, and the benefits are big.

Another reason was that Evil defaulted all
Magit buffers to Emacs mode. This was pain
while trying to learn and also broke all of
my habits for moving around. Now that I
fixed that by making all of those default to
motion state, that problem is nicely brushed
aside.

Finally, I had some concerns about some edge
cases being handled better by native Git CLI
tools - but for now I can keep the latter
available as a fallback.

Meanwhile, the big benefit is that the diff
display had a great built-in interface for
staging/unstaging/restoring - the rest is
less is less significant to me but the many
little niceties will probably add up as I
get used to them and discover new ones.