.emacs: add done message to pop-to-command-eshell
Super useful for distinguishing no-output cases,
such as `git diff` or `git stash list` with no
output.

Eat already provides something like this, so I
mostly followed Eat's lead for the format of
the message.

If `pop-to-command` became a public package, it
would be better for `pop-to-command-buffer` to
be globally `nil` and locally `t`, and to have
`pop-to-command-eat` set it too just like
`pop-to-command-eshell` does, but just doing a
quick-and-dirty version for my needs for now.