.emacs: make pop-to-command reset cursor position
If the cursor is not at the bottom of a reused
pop-to-command (Eshell variant) buffer, it will
stay in that position even after a new command
is started in the buffer. This is misleading at
worst, and at best not the most useful behavior.

One way this can happen is if I move somewhere
in the buffer after the command finishes, which
I do often.

Another way is this Eat bug:
https://codeberg.org/akib/emacs-eat/issues/60

This commit solves this problem by moving the
cursor to the end of the buffer before running
each command. (Only the Eshell variant needs
this - the Eat variant already resets the
cursor out-of-the-box.)