.emacs: basic git push and pull keybinds
Using repeated final keystroke as a kind of
typo-resistance, because doing a push when
you mean to pull could be unfortunate so.

No mnemonics for the letters, besides "t"
for "tags", "u" almost incidentally occurs
in "pull" (and "push", but "y" is closer to
"t" for tag push than "u" is, and so it
would make less sense to give "u" to push).