.emacs: more (un)selected color-coding
The modeline is already a good indicator, but
on larger screens a more full-window change is
helpful, and if the window selection changes
when I am not expecting it (examples: stray
touch on touchpad or touchscreen generating a
click event; bug in some code I wrote; keybind
is bound wrongly or not in effect in a mode I
try to use it in; command pops window but does
not select it), it's nice to have a visual
indication right where I happen to be looking
(for example, if I expect a window to remain
selected and suddenly it "dims").
remain selected).