.emacs: add trailing newline when creating files
Annoyingly, this also effects Eshell redirection
of program output to a file, since that actually
goes through Emacs machinery as if you created
the file the normal manual in-Emacs way. Mixed
feelings about this - in principle I hate it, in
practice I have often found it annoying when a
CLI command redirected to a file doesn't end in
a newline, so I guess we'll see if it annoys me
more than it helps. Either way, it's not super
obvious what a proper fix for this would look
like.