ssh-to-"identity" scripts
I wrote these years ago so that I could SSH
to devices whose IP addresses would change
without having to re-add them to the known
hosts file. By using a separate hosts file
for each device, with a wildcard address, I
was able to get the "allow this at any
address" behavior without the security risk.
In fact there was a small security benefit:
if I already know that I intend to SSH to
"bob", the only host key valid on that
connection is bob's.