.emacs: make ":q" in vi work sensibly
This has been a pain-point for me at least since I started
using Evil, if not earlier. How does it make sense for ":q"
to close the current window instead of the buffer in that
window!? (not even in addition to! just instead!)

Anyway, to me, it makes far more sense to have ":q" close
the buffer/file, because it's the inverse of opening a
file, and opening a file does not normally create a new
window. So why would I want my previously-set-up window
killed when doing my "close this file" muscle memory? And
why would I want the file to NOT close when doing CLOSE.
THE. FILE. muscle memory?

Maybe this makes slightly more sense inside Vim (it would
still be objectively worse and wronger but let's pretend
that I might be wrong on that point), but Emacs is less
like a vi-style editor and more like the shell, terminal
multiplexer, and one or more vi-style editors launched in
it. When I type ":q", that's far more like closing a vi
instance which just has that file open, than it is like
closing a pane/split/whatever-they-call-it inside a vim.

I kept putting up with it because I had other priorities,
although over the past month or two I glanced at how I'd
solve it for a few minutes here or there.

Finally fed up enough (and other priorities cleared) to
give it a more in-depth look, and luckily it didn't take
me very long to hone in on this solution (initially I
tried setting a different function in `evil-ex-commands`,
but advising `evil-quit` works more globally and reliably
with less lines of config/change - in particular, that
one advice also fixes ":wq" and doesn't break with-editor).