.emacs: try " zm" for media re-scan on Android
When photos are programmatically added to the
DCIM/Camera folder on Android (for example,
when copied by Syncthing to an Android device
from another device), and possibly other
folders, they don't show up in Google Photos
or photo pickers in other apps (like when
attaching images). Forcing a re-scan of the
shared storage partition/directory solves it.