.emacs: fix pop-up eshell history interference
I had originally removed history-related settings
from the Eshell variant of pop-to-command because
they seemed unnecessary, but it seems that if the
defaults are left alone then those Eshell buffers
can clobber the history by saving over it.

(The bigger bug is that Eshell history has no
handling for another Eshell instance concurrently
updating the history file. But that's a bigger
problem for me to solve later.)