.emacs: fix git pop-up working-directory edge-case
If `pop-to-command-eshell` is ran in a directory
like `~/a-git-repo/subdir`, then `vc-root-dir` is
`~/a-git-repo` but the wording directory of the
`git` process is `~/a-git-repo/subdir`. Normally,
this doesn't matter because `git` will do its own
search for the repo's root directory and behave
the same way. But it turns out that `vc-root-dir`
ignores Git's ceiling directories (or it bypasses
my `git` wrapper which sets the
`GIT_CEILING_DIRECTORIES` environment variable -
I haven't checked, but that seems less likely).
So if `~/a-git-repo` is a ceiling directory, then
a `git` command executed in `~/a-git-repo/subdir`
will act as if it is not in a Git repo.

In practice, this became a problem because I have
my home directory in Git, but I also have it as a
ceiling directory so that if I run a `git` command
in some directory which was meant to be a Git repo
but isn't, it errors out instead of silently doing
the wrong thing by operating on my home Git repo.

As a result, if I was in `~/code` and hit f.e. " vv"
for `git status`, and the `*eshell: git status (~/)*`
buffer didn't exist yet, that buffer would be
created (since `(vc-root-dir)` would return `~/`)
but it would run the `git` command in `~/code`,
which would respect the ceiling directory and
error out since it's "not" in a Git repo. And then
that buffer would stay in `~/code` and get reused
even if I re-ran the keybind/function from a buffer
that was in `~/`.

So the quick and minimal solution is to make my
pop-to-Git-command functions go into the
directory returned by `vc-root-dir`. The ideal
solution would also include making `vc-root-dir`
work in a way that reuses (or at least mimicks)
my `GIT_CEILING_DIRECTORY` setting. But even if
the latter is done, it's probably better to still
do the former as well (especially as long as these
git pop-up buffers explicitly include vc-root-dir
in their names and get shared/reused accordingly).