.emacs: eshell/vi: fix opening multiple relative paths
When my in-Eshell vi command was called with
multiple relative path arguments, if a later
argument (opened earlier, to get similar-to-vi
order in which I would see the files) was in
another directory, it would open and focus that
buffer before the next `find-file` would start,
so each `find-file` would run relative to the
next-in-arguments/previous-opened file's directory
instead of the directory from which the command
was started. So for example

    vi src/foo.c src/qux.c

would open src/qux.c, then try to open src/foo.c
relative to qux.c, so it was trying to open
src/src/foo.c instead.