.emacs: eshell/vo: more intuitive opening order
In eshell/vi, it made more sense to reverse the
order, since the buffers would display in most
recently opened order in one window.

In eshell/vo, it makes more sense that the first
argument will be opened in the same other-window
that it would get if it was the only argument.
In other words, it makes sense to mentally plan
which files will appear in what windows in the
order that they are in the argument list.