.gitconfig: make `git status` never mutate index
By default, "git status" (and some other commands?)
will opportunistically update the cached .git/index
information. The "--no-optional-locks" option will
cause Git to avoid locking that file, which is a
hilariously implementation-centric way of saying
that Git will not update that cached index file
(because it cannot, because it has been told to not
try to acquire the lock to safely write to it).

This might come back to bite me for repos with huge
unstaged (uncommitted?) changes, but in practice
the default behavior is already biting me sometimes
when I try to resume work on a code directory that
is supposed to be synced from another device with
something like Syncthing.