xdg-open for Windows: fix path handling
When I initially wrote this, it was for
URLs, so that worked fine without doing
any file path handling, and I think I
was aware of the problem for file paths
but just didn't care?

Then I forgot about this only supporting
URLs, and coincidentally it worked fine
for relative paths from either of the
WSL home directory or the Windows user
profile directory on files which have
identical relative paths in both, which
happens because I use symlinks from WSL
to Windows for folders like ~/Documents.

Sadly, we can't do the same handling for
both, but "does this exist in the file
system" is a good-enough heuristic for
now.