.emacs: dim the minibuffer harder
Every once in a while (best I've got is that
maybe it's somehow linked to opening/reverting
files with a newly installed mode for another
a programming language, because the last time
I distinctly remember it happening soon after
I installed csharp-mode and this time I just
recently installed rust-mode), the minibuffer
get emptied out in a way that the after-change
hook ignores.

We can manually reproduce the problem with

    (with-current-buffer " *Minibuf-0*"
        (let ((after-change-functions nil))
            (erase-buffer)))

which makes me think that something out there
is wiping out the minibuffer and deliberately
turning off the after-change-functions on it,
perhaps due to the idea that it's doing its
own internal programmatic changes which it
can cleanly keep to itself (it is obviously
wrong since it's erasing things that were in
the minibuffer before it started messing with
it).