.emacs: window state for bury-buffer
Not sure if I'll ever use this, especially now
that I have "Q" for `quit-other-window`, but
I've wanted it in the past so it seems worth a
try - this is a case where I'll find out if I
really want it by seeing if I end up using it.
I might prefer something more versatile (such
as doing whatever "q" in the buffer does,
instead of hard-coding `bury-buffer), but we'll
see about that as well - this is good enough
for now.