.emacs: add locked select to window state
Initially my conscious reason for this was
just for consistency with the other window
states (and some long-lingering idea that
it would help in practice/explaining/demos).

But it also seems like it might be useful
in combination with the commit made right
after this one which replaces window state
bury/etc sub-states with bury/etc actions.