.emacs: improve window state keybindings
Moved "n" back to "b". Mnemonic "bury", which
is more direct and memorable (I originally had
it on "b" but then moved it to "n" at the last
second... seemed like a better idea on full
desktop keyboard, but in practice I'm finding
"b" more intuitive here, plus "b" has the plus
of being more reachable by eitber hand.)

Buffer-killing actions (potentially destructive)
moved from "d" to "r", so that attempts to hit
"s" and "c" can't accidentally kill buffers.
"r" was one of my earlier ideas for this keybind
- it was only because "t" was bound next to it
that I didn't go with it. Mnemonic is "remove".

"t" is moved to "a" to clear area around "r".
(The mnemonic is bad here - the variable is
called "evil-auto-balance-windows", so it's a
stretch - "auto" is less relevant than "balance"
in this case. But bury/unbury gets priority,
both for mnemonic and ergonomic fit.)

(Assuming QWERTY. But better keyboard layouts
like Dvorak give less cause for typos on desktop
and laptop keyboards, and add far less value in
touch or thumb-typing keyboards, so proximity in
those layouts is less of a problem.)

I've yet to accidentally do this exact thing,
but accidentally hitting nearby buttons is a
common-enough error. Emacs usually protects
buffers from killing if they have unsaved file
changes or running processes in them, but that
won't protect against accidentally wiping out
every buffer that might hold useful work/state.

Not using modifier keys as the safety because
I want to alternate fluidly between bury and
bury-to-kill, and I don't want the overhead of
an extra keystroke in a workflow that I could
see becoming very common for me.

I actually kinda want buffer bury and buffer
kill actions next to each other, but the risk
of adjacent-key typos is holding me back.