.emacs: hack-avy-tree: fix advice removal edge case
I was originally trying to be extremely careful
against even possibilities like `avy-tree` being
called by some other code before `aw-select` got
around to calling it. (To the best of my memory
anyway, turns out I failed to record that in my
commit comment.)

Of course, this means the workaround would fail
to remove the advice in any case other than the
single-window case until after `aw-select` is
done, which means the advice would still be on
the function during the callback of `aw-select`
that actually acts on a window (which is by far
the most realistic case for calling `avy-tree`
a second time during a call to `aw-select`).