.emacs: window state: "gp" and "gP" shorthands
I find myself wanting this in the common case of
copy-pasting windows. Putting windows in registers
is nice, but in window state v1 I could just do
" oc2" to copy the current window to window two. In
window state v2, this has become "syysp2s2" or
"Syyp2{{Esc}}".

Now I can do "sgP2" (or "sgp2" if I want to "paste"
but keep the current window selected instead of
selecting the "pasted" window).

Limitations of avy/ace-window as a base are really
showing here in what I had to do to implement "g"
as a prefix keybinding in aw-window-dispatch.