.emacs: fix duplicates in consult-line history
There's something weird going on with the history,
something adds a duplicate and I'm not sure when
or where (I suspect async code has something to
do with it, since the duplicate exists after the
`delete-dups` in `consult-line-quit`, but another
duplicate deletion right before resuming a search
at least keeps it from messing up the indexing in
my UX (f.e. "4 L" in vi gets the right thing, etc).