.emacs: remove Termux clipboard hack
Haven't been using this for months now because the
battery drain of calling `termux-clipboard-get` in
a loop regularly was too high. I was holding on to
it in the hopes of a faster and more efficient API
solving that problem (the clipboard getting itself
is basically instant - it's the stupid design of a
whole Dalvik VM cold start every time termux-api-*
programs run) but I haven't and won't have time to
tackle that any time soon.