.emacs: add buffer-differs-from-visited-file-p
Emacs doesn't have an API for it, but it has the
functionality built-in inside `set-buffer-modified`,

`set-buffer-modified-p` will **sometimes** check if
file contents match buffer contents, as part of
deciding whether to do that `{{file name}} changed
on disk; really edit buffer?` prompt which you'll
sometimes see in the minibuffer. But it only does
that check if

1. the call is setting the modified flag to true,
2. the buffer's modified flag is currently cleared
   (Emacs believes the buffer is not modified), and
3. the buffer's last-modified time is more recent
   than what Emacs currently remembers as the
   underlying file's last-modified time.

(None of this seems to be documented or guaranteed
as a stable feature of the API. In fact I would've
never expected `set-modified-flag-p` to also diff
file and buffer and then fire off a call to another
user-customizable function as a secret conditional
side-effect testing the behavior. I only found out
because someone told me to check those functions,
which seemed like they had less of an idea than I
did, but I check before calling people wrong when
I can, and what a pleasant surprise that was. And
we'd have to look in the C source to rule out
`set-buffer-modified-p` doing something else given
some config or other. But it seems to work, and
seems reasonably likely to remain stable.)

So what we have to do is basically three tricks:

1. clear the modified flag,

2. set the cached "when was the underlying file
   last modified?" value far enough back in time
   (I assume `'(0 0)` is either the minimum or at
   least will always be low enough).

3. run `(set-buffer-modified-p t)` with advice on
   `ask-user-about-supersession-threat` to detect
   if the prompt would've triggered.