.emacs: improve window state yank/paste
1. Window state registers are now separate
   from normal vi/Emacs registers. (When
   is it useful to have them clobber each
   other? When am I going to use the stuff
   in my window state registers as normal
   Emacs registers? or for a normal paste
   operation? And so on. It's not a useful
   sharing/overlap for me so far, but it
   is useful to have them separate so that
   I don't clobber text copy-paste while I
   do window copy-paste and vice-versa.)
2. Proper default registers: 1-9 as a ring
   of the last nine yanks or deletes, 0 as
   the last yank (and '"' as the "unnamed"
   register - I copied Evil behavior of it
   just being an alias for "1").
3. Fast paste no longer goes through Emacs
   registers either: it only made sense to
   do that because the other yank/paste
   code was using Emacs registers and the
   code was simpler that way.

I matched Evil's behavior where explicitly
giving a default register for a yank or
delete is just ignored (even though to me
it makes sense that f.e. `"8yy` should
override the 8th entry in the ring), since
I want to stay consistent w/ Evil where it
doesn't seem too unreasonable. But I didn't
copy Evil's quirk of not setting register 0
for a yank in that case, because that seems
less like an intentional thing and more
like a side-effect of a combination of
conditional branches written without this
edge case in mind.

I did not match the Evil behavior of yanks
or deletes to named registers also pushing
an entry to the unnamed ring - because

1. the whole point of the unnamed ring in
   my mind is to have access to the copies
   you didn't name (so clobbering unnamed
   copies with named copies is redundant
   and needlessly reduces your capacity)
   (but I recognize that I could be missing
   some wisdom here - neovim also pushes to
   the unnamed ring when copying to named
   registers), and
2. it doesn't seem to be consistent with
   vi/vim/neovim - it seems to be an Evil
   quirk, I'd guess one that made sense
   for integration with the Emacs kill-ring
   and system clipboards.