.emacs: fix histdir load order
This will break (non-data-destroyingly)
once the histdir has more entries than
the ring size: it will load the oldest
entries, not the newest.

This is all assuming that the raw order
of file listing from the OS is based on
the time of creation/modification, an
assumption which in my experience is
traditionally true on UNIX-like systems
and modern Linux, but is not guaranteed
and might be file-system dependent as
far as I know. (In fact, it seems only
loosely true on my systems, although
that might be Syncthing-related - which
is great by the way, brings much joy to
have synced history across my devices!)

In six months I've ended up with <2000
history entries. So in a decade or four
from now I might finally hit 65535, and
by then everything will be so different
that it's unlikely that solving it now
will pay off. Even this much words and
thought paid it is giving it far more
than it is worth, but I needed to think
it through to "break free" and I hope
writing it down will help someone else.