.emacs: Eshell: normal shell up/down
Thought I'd get used to it, but no. Or at least
not universally. I finally started to naturally
reach for it in some cases, but too often I hit
up after breaking the "up" streak, which resets
the history cycling to use the current input,
and that is never what I want.

Perhaps if it could only do matching next/prev
when the input line was typed in its entirety,
but I have no idea how to do that and it's not
worth the effort when I will be happy enough
with just normal up/down (I have my "gh" for
consult-history as a better version of what the
default up/down Eshell keybinds do anyway).