.emacs: fix diff-unsaved-changes header
This reverts commit 46cda7eac3f530dd08202a0ef79f9fddc7bd1a06

Also the comments in that commit were wrong,
I just forgot: the `cdexec ... git diff`
instead of `gd` was ensuring that the text
at the top of the diff said "file" and
"unsaved" instead of f.e. "tmp/18fh3d/file"
and "tmp/18fh3d/unsaved".

Now that my `gd` is updated to use `--no-index`,
it can be used in place of `git diff` (because
it won't `cd /` anymore), but both still need
to be ran in the temporary directory; might as
well cut out the dependency entirely. Speaking
of `--no-index`, might as well add it here for
total edge-case robustness.