.emacs: abomination, saved in the Git record
This was an attempt to hack together a comint-style
interface on top of Eat, using Eat's Emacs mode to
hold off on sending output to the terminal, and a
couple keybindings (Enter and Tab) to manually send
the inputs. I don't expect this to work in the
general case, but it seems like it should work for
most or even all REPL use-cases I've ever seen.

However, it's very fragile. It relies on several
assumptions about how Eat is implemented, and it
relies on Tab in the REPL never causing the prompt
of the REPL to change length.

It works great in Python. It seems to only partly
work in Node.js, but I've already spent far too
much time on it.