.emacs: wrap any REPL in Eat with histdir (part 2)
Make consult-history run in a temporary buffer.

The price: no dynamic preview in the buffer of
the history candidate currently selected. I am
not immediately sure how much I'll miss this.
Probably more than I realize, because it seems
that I'm already noticing the annoying loss of
UX efficiency in some situations (I think the
biggest two factors are visual distinctiveness
and more importantly the fixed location - the
command you would currently run showing up in
the Eshell/comint/etc input as you search is
nice because it doesn't require looking over
to where the search is, and it's always in the
location where I was already looking).

But this turned out to be necessary for REPLs
in a terminal because of the myriad possible
ways a REPL could trigger a terminal redraw
and thus mess up the interaction between the
consult-history preview and what Eat was doing,
and because consult-history does its own thing
to render the preview (... or was it necessary?
It's possible that we could reimplement the
preview in a way that works with REPLs in Eat
by getting more invasive with advice, but I
have already given this a lot of time so I
won't be pursuing that unless the lack of the
redundant preview ends up being a UX problem).