.emacs: fix histdir-repl-{beginning/end}-of-line
Improves reliability of the cursor position being
consistent between REPL/terminal and Eat/buffer,
by sending a character and backspacing over it,
which should force a redraw in more (all, unless
the REPL or terminal is trying to be clever with
optimizing out some character echoes).

This should reduce (eliminate entirely, I suspect
and hope) the need for the `(end-of-buffer)` call
as a hack to get the cursor to align when going
into insert state.