.emacs: histdir-repl: rename "line" to "input"
I initially went for "line" because this whole
thing only works if we assume REPL input is a
single line.

But the more this grew, the more I kept seeing
that "input" would better match (and convey or
self-describe) the semantics. So I've been
intending to do this renaming for a while.

And now I have good reasons to reuse a couple
of the `histdir-*-line*` names.