.emacs: histdir-repl: vi-style h/l/0/$ motions!
Crucially, these Just Work and Do The Right Thing
whether point is in the REPL input or not.

The 0.04 sleep is a little annoying and this is
perceptibly slow (and sometimes seems to lock up)
when I hit "h" or "l" too many times too quickly,
but it's a start.

I am starting to realize that the correct hook is
actually more general: after every motion, or on
a timer or when idle (idle is probably the right
move), if point is in REPL input and not matching
the terminal's cursor position, send one batch of
left/right arrow key escape sequences to align it.
But that'll have to be a later commit.