.emacs: histdir-repl: full vi replace state!
This will break on unforeseen reasons for the
input line to grow/shrink, unfortunately. In
principle we could fix by re-"entering" input
in more situations, but that is disruptive and
has to wait for cursor-moving operations to
round-trip through the terminal - in principle
that is the only "correct" way; in practice I
see the slowdown when holding down a key that
does that on each keypress (my earliest "h"/"l"
keybinds for histdir-repl did that) (but it may
be possible to stay reliable while reducing the
sleeps by an order of magnitude or more).