.emacs: histdir-repl vi "o" and "O" as "A" and "I"
I was undecided about what to do with these for a
while.. the vi behavior of o/O is inherently
multi-line, whereas I'm keeping the REPL wrapping
single-line to keep things manageable (and almost
every REPL/shell can only do done line of editing
at a time anyway... off the top of my head, only
zsh's ZLE can do multiple lines, and even that is
only pratically accessible in its vi mode).

One possibility I was thinking of was to have "o"
basically be "\ni" - send the current input and
start a new line, and to have "O" copy the input
into a stack, and enter insert mode for the next
input, and so long as that stack has items, each
new input would be initialized with the contents
of a popped stack entry. These seems sorta useful,
but not useful enough. "O" would be redundant vs
the "one obvious way to do it", which is more
flexible/powerful anyway (namely, just "dd" or
"cc" the input and you have a stack in the form
of numbered registers, or a map in the form of
named registers). Also, "o" would be... badly
close to much safer keys which you might hit
precisely to edit input that you absolutely don't
want to run as-is (right next to "i" on QWERTY;
right next to "a" on Dvorak).

So that leaves either unbinding the o/O keys and
leaving them for some other reuse later, or give
them the closest single-line equivalent to their
normal behavior - and the latter is same as A/I.