.emacs: stop having Escape as alias for quit/bury
Escape made sense as a quit/bury key for buffers
which don't contain arbitrary things which may
need an arbitrary amount of Escape keystrokes.

REPLs, on the other hand, sometimes take Escape,
and more importantly, I reflexively think REPLs
might understand Escape for escaping out of some
UI elements, like irb's completion candidate TUI.

So it's very natural for me to hit Escape one or
more times just as a "quit/abort/go-away" reflex
at something in the REPL, while not wanting the
REPL itself to go away.

In other words, I am finding myself regularly
minimizing the REPL buffer/window when I didn't
mean to, and I really don't like it.

TL;DR: lesson: in a modal interface, if a thing
doesn't take/process your escape key at all (and
that includes anything that might sometimes be
nested in that thing), it is good to reuse that
escape key for quitting or minimizing it; but if
it or its contents do use the escape key, don't
reuse the escape key for quitting or minimizing
that thing.