.emacs: UI/UX polish in help/etc
Since I have Ctrl-Space as a more universally
bound alias for Space, and on terminals that's
Ctrl-@ (it's an ASCII thing), and Ctrl-@ sorts
before Space (since Ctrl-@ is actually a nul
byte... it's an ASCII thing), Emacs has been
showing all of my space keybinds as bound to
`C-@ k` instead of `SPC k`. This finally fixes
that.