.emacs: dlist: lisp-friendly doubly-linked list
Big benefit TL;DR: the cdr of a dlist is a lisp
list (the normal singly-linked kind) of all the
elements in the dlist. This allows the dlist to
still be used at zero cost in any non-mutating
function that takes a list.

This is achieved at the cost of two additional
cons cells per element than a regular list, or
just one cons cell more than the obvious naive
doubly-linked list implementation. We create a
normal list, then a wrapper cons cell for each
cell of the list whose cdr is the list's cell
and whose car is two links to adjacent wrapper
cells. For example, `(dlist "A" "B" "C")` has
cons cells like this:

(nil  +-)--    --(-+   +-)--    --(-+  nil)
           \  /             \  /
  ^         \/     ^         \/     ^
  |         /\     |         /\     |
( +   + ) <-  -> ( +   + ) <-  -> ( +   + )
      |                |                |
      v                v                v

("A"  +-)------> ("B"  +-)------> ("C" nil)

One quirk of this structure is that we have two
links tailwards: one in the normal list and one
in the doubly-linked link conses, but the former
is needed to provide a normal lisp list, and the
latter is needed when adding or removing items
(and for any tailwards traversal that wants the
option to reverse direction).

For naming:

* dlist because it's not just a doubly-linked
  list (so `doubly-linked-list` would be lying),
  because it's got multiple angles of mnemonic
  potential ("doubly"; "D-list" self-deprecating
  joke about obscurity; very similar to "list").
* `car`, `cdr`, and `nth` names mirror the big
  picture semantics and usage shape of their
  standard list counterparts.

Just the minimum set of functions needed to do
useful things for now. For example, I left out
a separate link function because that can be
done with a combination of cdr, cons, and setcdr
just as with normal lisp lists; but I included
`dlist-unlink` because that has to take care of
dlist-specific nil caes when updating the links;
and I included `dlist-nthcdr` because it was in
the fastest and easiest path towards testing
`dlist-unlink`.