.emacs: implement ordered hash table
Big benefit TL;DR: the cddr (cdr of cdr) of the
ordered hash table is a list of all elements in
the hash table, from most recently inserted to
least.

This implementation uses dlist, and as a result
the list is available "for free" (at O(1) cost).

This is useful for integrating with all sorts
of existing code. For example, consult-history
and completing-read both lack any practical way
to plug in an arbitrary data structure, so an
implementation which provides a native list at
tiny O(1) cost is great for switching history
mechanisms like histdir over to an ordered hash
table data structure.