.emacs: fix fixed-consult-history erasing buffer
If a buffer can't have history (for example, when
just editing a normal file in Fundamental Mode),
then my `fixed-consult-history` was calling
`delete-field` (nowadays through `delete-command`)
before the error from not having history.

This is normally a non-issue, but it's a little
annoying to have a split second of confusion (and
fear of having lost work) and then undo when I
accidentally hit my "gh" or "gr" keybinds in the
wrong buffer.

When I first created this problem, I wasn't calling
`consult--current-history` directly (trying to keep
the private interface use to a minimum),  but now
that I am, the fix is a simple reordering of the
delete after that call.