.emacs: workaround Eshell process sentinel issue
When I switched from having `diff-unsaved-changes`
change directories through `cdexec` to having it
set Emacs' default-directory, it created an issue
where some "process sentinel" which remembers the
default-directory from the previous run errors out
when pop-to-command-eshell reuses the buffer.

Making the external process do the directory
change on its own is the simplest and most obvious
way to fix that.