.emacs: work around Eat end-of-buffer property issue
Eat 0.9 added a `rear-nonsticky t` property:

https://codeberg.org/akib/emacs-eat/issues/92

Usually rear-nonsticky is the right move, but in
this specific case it doesn't work as the default
because it messes up motions when point is at the
end of buffer.

I'm not sure if a global hook will mess anything
up in some edge case, but we'll let it ride.