.emacs: lambda-let
Many many times I have wanted to write a function
and have it bind some variables at the same time.

In many ways this is a stopgap/workaround for not
using lexical scope in my .emacs - I only ever
want this due to stuff not lexically binding.

This will cover some cases of that - not all, but
it covers cases where I only need fully-bound
(as opposed to partially bound) lambdas.

Another way of looking at it: `lambda-let` is
turning a `let` form into a function.

`varlist` is so named because that's the same
name used by `let`.