.emacs: lambda-partial-let
Generalization of lambda-let: a lambda where only
some of the arguments are bound by a `let`-style
varlist (`lambda-partial-let` is to `lambda-let`
as `apply-partially` is to `apply`/`funcall`).

The `args` parameter (same name as in `lambda`)
is given after the `varlist` parameter and names
all the parameters that are not bound.