.emacs: perfectly responsive histdir load
Almost as good as it could get even in principle.
The most recent calls are loaded first and are
immediately available as history while the rest
of the older calls keep getting read in in the
background.

History cycling still needs to be reimplemented
(up/down arrows) and I'd like to add support for
local history (per buffer) which is not changed
by sync and always sorts before synced history
(default one local most recent entry, settable
to more with "gl" to select an entry and then
that's how many extra most-recent local entries
are specially kept).