.emacs: cleaner ordered hash table design
Don't make it a dlist whose head is a hash table;
make it a cons cell whose car is a hash table.

This was what I was going to do at first, but then
it seemed like a waste to have extra logic for
removing the most recent entry when dlist-unlink
could just handle it if only the whole thing was
one dlist.

The problem is that unlike a list, a dlist has a
link headwards, and so usage of the ordered table
that wants to work with the entries as a dlist
ends up needing to know the implementation detail
that the entry dlist's head is actually headed by
the ordered table itself.