Revert ".emacs: input and inputter helpers for unread-command-events"
This reverts commit d8e59512c397f2624da4c59d8199cd403001bd16.

Every so often I find myself wanting to add these,
and then end up realizing that I don't actually
have enough use for them (I do have some uses of
`(setq unread-command-events ...)` elsewhere in
the code, but it's not entirely clear that those
are worth switching over to these helpers just yet.

So erring on the side of YAGNI/wait-and-see yet again.