.emacs: add read-key-sequence-in-keymap helper
Yet another piece that I have wanted a few times,
although now that I've figured it out and tried
using it (to dispatch space-prefixed commands to
other windows through my window state) I doubt
if it's actually the right/best way to do any
particular solution.

Implementation notes: setting `overriding-local-map`
causes almost all possible keymaps to be ignored,
and in particular that's important so that we do
not end up searching keymaps specified in the
properties of text under point (my first thought
to avoid this was to wrap the key sequence read in
`(with-temp-buffer ...)`, but that caused wrong
behavior when the key pressed was an Escape. That
just leaves `overriding-terminal-local-map` and the
global map to take care of.