.emacs: fix " q" and Space->Escape keybinds
I seem to have sloppily defined them only in
normal mode (discovered this a couple days
ago as part of testing `read-key-sequence-in-map`
with space-map in the hopes of adding a window
state operator to do my space-prefixed commands
in windows other than the focused one).

Also, I keep putting off actually doing it, but
this is probably the clearest demonstration of
why I should just be defining these keys through
the space-map variable itself and not indirectly
through evil-*-state-map.