.emacs: window state: " " runs a space-prefixed command in another window
I wanted this for some time, but it was a little
too far out-of-reach (and maybe not obviously-enough
valuable-enough) until recently.

However, implementing this has given me more reason
to think that I should probably implement a general
ability to dispatch any keystroke(s) to a window,
or perhaps to switch back to the most recently used
window, or both. In particular, I still want/need

1. the ability to send "q" to a window, as a more
   deterministic/reliable form of my current "Q"
   binding (perhaps also Escape), and
2. the ability to start or switch to one of several
   Eshell instances - for example, using universal
   prefix numbers or vi register letters to refer
   to specific instances.

Both of these might be worthy of their own keys in
window state ("t" is free for Eshell, mnemonic for
"terminal" and matching how I already use " t"; it
is perhaps worth sacrificing "q" as a window-state
quitting key for sending "q" to other windows), but
it would be nicer to also have a general-purpose
key to do any command, for maximum flexibility).

(Also, keymaps can contain bindings that are not
commands, and my current solution doesn't support
any bindings that `call-interactively` can't call.)

But for now, a space-prefixed keybind dispatcher
is a good incremental step in that direction.